slam-rs: split IMU and estimator modules by responsibility
