+++ packages/slam-rs/crates/slam-rs/src/estimator/mod.rs
@@ -29,22 +29,25 @@
 mod frame_update;
 mod optimize;
 mod schedule;
+mod snapshot;
+mod triangulation;
+
+pub use snapshot::{SnapshotLandmark, VelBias, WindowSnapshot, WindowState};
 
 use std::collections::{BTreeMap, BTreeSet, VecDeque};
 use std::sync::Arc;
 
-use nalgebra::{DMatrix, DVector, Vector2, Vector3, Vector4};
+use nalgebra::{DMatrix, DVector, Vector2, Vector3};
 
-use crate::ba_base::{BaError, BundleAdjustmentBase, triangulate};
+use crate::ba_base::{BaError, BundleAdjustmentBase};
 use crate::calib::Calibration;
-use crate::camera::CameraEnum;
 use crate::config::{LinearizationType, VioConfig};
 use crate::duration_ns;
 use crate::imu::{
     ImuError, ImuLinData, ImuNoise, ImuSample, IntegratedImuMeasurement, Popped, gravity,
     gravity_from_first_accel,
 };
-use crate::landmark::{Landmark, LandmarkError, StereographicParam};
+use crate::landmark::LandmarkError;
 use crate::lie::{LieScalar, Se3, eigen_maxi};
 use crate::linearize::LinearizeError;
 use crate::marg::MargError;
@@ -364,74 +367,6 @@ pub struct FrameStats<S: LieScalar> {
     pub frame_update: FrameUpdateOutcome,
     /// Wall-clock stage marks; see [`StageTimings`].
     pub timings: StageTimings,
-}
-
-/// The nine degrees of freedom a 15-dof state carries beyond its pose.
-///
-/// A pose block (`frame_poses`) has none of them, a state block
-/// (`frame_states`) has all three, so they travel as one value rather than as
-/// three `Option`s that are always all-`Some` or all-`None`.
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct VelBias<S: LieScalar> {
-    /// World-frame velocity.
-    pub vel_w_i: Vector3<S>,
-    /// Gyroscope bias.
-    pub bias_gyro: Vector3<S>,
-    /// Accelerometer bias.
-    pub bias_accel: Vector3<S>,
-}
-
-/// One window state, as the S9 Rerun rung needs it.
-#[derive(Debug, Clone, PartialEq)]
-pub struct WindowState<S: LieScalar> {
-    /// State timestamp.
-    pub t_ns: i64,
-    /// `T_w_i`, the rig pose in the world frame.
-    pub t_w_i: Se3<S>,
-    /// The nine dof beyond the pose, `None` for a pose-only block.
-    pub vel_bias: Option<VelBias<S>>,
-    /// Whether the linearization point is frozen.
-    pub linearized: bool,
-    /// Whether this frame is a keyframe.
-    pub keyframe: bool,
-    /// Whether this frame is a long-term keyframe.
-    pub long_term_keyframe: bool,
-    /// Monotonic frame index for the UI.
-    pub frame_index: Option<usize>,
-}
-
-/// One landmark as the V2 rung draws it.
-///
-/// The host is carried because the rung colours the point cloud by the keyframe
-/// that hosts it: the position alone cannot say which frame's bearing it is a
-/// distance along, and `lmdb` keys on the host rather than storing it per point.
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct SnapshotLandmark<S: LieScalar> {
-    /// The landmark's id, which is the id of the keypoint that spawned it.
-    pub id: LandmarkId,
-    /// Host keyframe and camera: the image the inverse distance is measured from.
-    pub host: TimeCamId,
-    /// Position in the world frame, metres.
-    pub position_w: Vector3<S>,
-}
-
-/// The window and its landmarks, for the V2 visual-validation rung (D51).
-///
-/// `getAllPosesMap`, `get_current_points` and the `VioVisualizationData` fields
-///  collapsed into one value the caller reads once per frame. The
-/// core logs nothing itself (D03).
-#[derive(Debug, Clone, PartialEq)]
-pub struct WindowSnapshot<S: LieScalar> {
-    /// Frameset timestamp of the newest state.
-    pub t_ns: i64,
-    /// The 15-dof states, oldest first.
-    pub states: Vec<WindowState<S>>,
-    /// The pose-only blocks, oldest first.
-    pub poses: Vec<WindowState<S>>,
-    /// Landmarks the window currently holds, in `lmdb` order.
-    pub landmarks: Vec<SnapshotLandmark<S>>,
-    /// Frames the last marginalization removed from the window.
-    pub marginalized: Vec<FrameId>,
 }
 
 /// LM damping, the four fields of in one place.
@@ -852,79 +787,6 @@ impl<S: LieScalar> SqrtKeypointVio<S> {
         &self.imu_meas
     }
 
-    /// The window, its landmarks in the world frame and the frames the last
-    /// marginalization removed, for the V2 Rerun rung (D51).
-    ///
-    /// The landmark position is `host_pose * T_i_c * unproject(direction) /
-    /// inv_dist`, exactly as the landmark bundle is built, and a
-    /// landmark whose host has left the window is skipped as it is there.
-    pub fn snapshot(&self) -> WindowSnapshot<S> {
-        let states: Vec<WindowState<S>> = self
-            .ba
-            .frame_states
-            .iter()
-            .map(|(t_ns, state)| {
-                let inner: &PoseVelBiasState<S> = state.state();
-                WindowState {
-                    t_ns: *t_ns,
-                    t_w_i: inner.t_w_i,
-                    vel_bias: Some(VelBias {
-                        vel_w_i: inner.vel_w_i,
-                        bias_gyro: inner.bias_gyro,
-                        bias_accel: inner.bias_accel,
-                    }),
-                    linearized: state.is_linearized(),
-                    keyframe: self.kf_ids.contains(t_ns),
-                    long_term_keyframe: self.ltkfs.contains(t_ns),
-                    frame_index: self.frame_idx.get(t_ns).copied(),
-                }
-            })
-            .collect();
-        let poses: Vec<WindowState<S>> = self
-            .ba
-            .frame_poses
-            .iter()
-            .map(|(t_ns, pose)| WindowState {
-                t_ns: *t_ns,
-                t_w_i: *pose.pose(),
-                vel_bias: None,
-                linearized: pose.is_linearized(),
-                keyframe: self.kf_ids.contains(t_ns),
-                long_term_keyframe: self.ltkfs.contains(t_ns),
-                frame_index: self.frame_idx.get(t_ns).copied(),
-            })
-            .collect();
-
-        let mut landmarks: Vec<SnapshotLandmark<S>> =
-            Vec::with_capacity(self.ba.lmdb.num_landmarks());
-        for lm in self.ba.lmdb.landmarks() {
-            let Ok(host) = self.ba.get_pose_state_with_lin(lm.host_kf_id.frame_id) else {
-                continue;
-            };
-            let Some(t_i_c) = self.ba.calib.t_i_c.get(lm.host_kf_id.cam_id) else {
-                continue;
-            };
-            let t_w_c: Se3<S> = *host.pose() * *t_i_c;
-            let bearing: Vector4<S> = StereographicParam::unproject(&lm.direction);
-            let scale: S = S::one() / lm.inv_dist;
-            let point_c: Vector3<S> =
-                Vector3::new(bearing[0] * scale, bearing[1] * scale, bearing[2] * scale);
-            landmarks.push(SnapshotLandmark {
-                id: lm.id,
-                host: lm.host_kf_id,
-                position_w: t_w_c * point_c,
-            });
-        }
-
-        WindowSnapshot {
-            t_ns: self.last_state_t_ns,
-            states,
-            poses,
-            landmarks,
-            marginalized: self.last_marginalized.clone(),
-        }
-    }
-
     /// Process one frameset synchronously.
     /// Initialize from the first accelerometer sample if needed, then preintegrate
     /// `(prev_t, curr_t]` using the previous state's biases and call `measure`.
@@ -1252,135 +1114,6 @@ impl<S: LieScalar> SqrtKeypointVio<S> {
             frame_update,
             timings,
         })
-    }
-
-    /// Triangulate unconnected observations into landmarks hosted by this frameset.
-    ///
-    /// Gather all observations of each id in the live window and try second views
-    /// in `TimeCamId` order. Unproject both pixels, form
-    /// `T_0_1 = T_i_c[host]⁻¹ · T_i0_i1 · T_i_c[other]`, reject short baselines,
-    /// and accept finite DLT results only when `0 < inv_dist < 3`.
-    /// On acceptance, file all gathered observations, not just the successful pair.
-    /// Each camera may host landmarks.
-    ///
-    /// All camera ids have been validated against the rig. Host pose and calibration
-    /// stay constant throughout the loop, so their inverses are computed once.
-    fn triangulate_unconnected(
-        &mut self,
-        frame: &FlowObservations,
-        unconnected_obs: &[BTreeSet<KeypointId>],
-    ) -> Result<usize, EstimatorError> {
-        debug_assert_eq!(unconnected_obs.len(), self.ba.calib.t_i_c.len());
-        // the squared threshold is formed in `double` and cast, so the
-        // `f32` instantiation compares against `(float)(0.05 * 0.05)`.
-        let min_triang_distance2: S = S::from_literal(
-            self.config.vio_min_triangulation_dist * self.config.vio_min_triangulation_dist,
-        );
-        let mut num_points_added: usize = 0;
-        // 's `T_i0_inv`: the host is this frameset, for every landmark
-        // and every pair.
-        let t_i0_inv: Se3<S> = self
-            .ba
-            .get_pose_state_with_lin(frame.t_ns)?
-            .pose()
-            .inverse();
-
-        for (cam_id, ids) in unconnected_obs.iter().enumerate() {
-            let tcidl: TimeCamId = TimeCamId::new(frame.t_ns, cam_id);
-            let host_keypoints: &BTreeMap<KeypointId, Vector2<f32>> = &frame.cameras[cam_id];
-            let cam0: CameraEnum<S> = self.ba.cameras()[cam_id];
-            let t_i_c0_inv: Se3<S> = self.ba.calib.t_i_c[cam_id].inverse();
-            for kpt_id in ids {
-                let lm_id: LandmarkId = LandmarkId::from(*kpt_id);
-                // another camera of this frameset may have hosted it
-                // already.
-                if self.ba.lmdb.landmark_exists(lm_id) {
-                    continue;
-                }
-                // 's `.at(lm_id)`: `measure` took this id out of
-                // `host_keypoints` itself, so a miss is an invariant break, not
-                // a landmark to skip (D32).
-                let Some(p0_pixel) = host_keypoints.get(kpt_id) else {
-                    return Err(EstimatorError::UnconnectedKeypointMissing {
-                        cam_id,
-                        kpt_id: *kpt_id,
-                    });
-                };
-                let p0: Vector2<S> = cast_pixel::<S>(p0_pixel);
-
-                // Visit every image of this id in the live window in `TimeCamId` order.
-                let mut kp_obs: BTreeMap<TimeCamId, Vector2<S>> = BTreeMap::new();
-                for (other_t_ns, other) in &self.prev_opt_flow_res {
-                    for (other_cam, keypoints) in other.cameras.iter().enumerate() {
-                        if let Some(pixel) = keypoints.get(kpt_id) {
-                            kp_obs.insert(
-                                TimeCamId::new(*other_t_ns, other_cam),
-                                cast_pixel::<S>(pixel),
-                            );
-                        }
-                    }
-                }
-
-                let mut accepted: Option<Landmark<S>> = None;
-                for (tcido, p1) in &kp_obs {
-                    // an unprojection the camera rejects skips this
-                    // pair, not the landmark.
-                    let mut p0_3d: Vector4<S> = Vector4::zeros();
-                    let mut p1_3d: Vector4<S> = Vector4::zeros();
-                    let cam1: CameraEnum<S> = self.ba.cameras()[tcido.cam_id];
-                    let valid0: bool = cam0.unproject(&p0, &mut p0_3d);
-                    let valid1: bool = cam1.unproject(p1, &mut p1_3d);
-                    if !valid0 || !valid1 {
-                        continue;
-                    }
-
-                    let other_pose: Se3<S> =
-                        *self.ba.get_pose_state_with_lin(tcido.frame_id)?.pose();
-                    let t_i0_i1: Se3<S> = t_i0_inv * other_pose;
-                    let t_0_1: Se3<S> = t_i_c0_inv * t_i0_i1 * self.ba.calib.t_i_c[tcido.cam_id];
-
-                    // Require enough squared translation baseline to triangulate.
-                    let t: Vector3<S> = t_0_1.translation;
-                    let baseline2: S = t[0] * t[0] + t[1] * t[1] + t[2] * t[2];
-                    if baseline2 < min_triang_distance2 {
-                        continue;
-                    }
-
-                    // A refused DLT skips this observation pair.
-                    let Some(triangulated): Option<Vector4<S>> = triangulate(
-                        &Vector3::new(p0_3d[0], p0_3d[1], p0_3d[2]),
-                        &Vector3::new(p1_3d[0], p1_3d[1], p1_3d[2]),
-                        &t_0_1,
-                    ) else {
-                        continue;
-                    };
-                    let finite: bool = triangulated.iter().all(|v| v.is_finite());
-                    // What decides this gate is the triangulated value's own
-                    // reduction order (D47), not the comparison.
-                    let inv_dist: S = triangulated[3];
-                    // `3.0` is a `double` literal, so the `f32` instantiation
-                    // promotes and compares in `double`.
-                    if finite && inv_dist > S::zero() && inv_dist.to_f64() < 3.0 {
-                        accepted = Some(Landmark::new(
-                            lm_id,
-                            tcidl,
-                            StereographicParam::project(&triangulated),
-                            inv_dist,
-                        ));
-                        break;
-                    }
-                }
-
-                if let Some(landmark) = accepted {
-                    self.ba.lmdb.add_landmark(lm_id, &landmark);
-                    num_points_added += 1;
-                    for (tcido, pixel) in &kp_obs {
-                        self.ba.lmdb.add_observation(*tcido, lm_id, *pixel)?;
-                    }
-                }
-            }
-        }
-        Ok(num_points_added)
     }
 
     /// `ImuLinData` as and build it.
